Benchmark instantiated CUTLASS GEMMs and trim PCH experiment artifacts
